fix: rollback mysql on OperationalError
The bot's been struggling in production environment because of Mysql
connection dropping during queries in some situations. This triggers
consistency errors in sqlalchemy because there's no rollback when the
error happens. This commit will rollback if any OperationalError is
raised and then it will raise the same error to fail the bot command.